added typescript to expo project
