+++ components/AlarmNameInput.js
@@ -5,13 +5,13 @@ import { getDynamicStyles } from '../styles/AlarmStyles';
 
 function AlarmNameInput({ alarmName, setAlarmName }) {
   const { isDarkMode } = useDarkMode();
-  const dynamicStyles = getDynamicStyles(isDarkMode); 
+  const dynamicStyles = getDynamicStyles(isDarkMode);
 
   return (
-    <View style={dynamicStyles.container}> 
+    <View style={dynamicStyles.container}>
       <Text style={dynamicStyles.label}>Label</Text>
       <TextInput
-        style={dynamicStyles.textInput} // Apply the dynamic styles here too 
+        style={dynamicStyles.textInput} // Apply the dynamic styles here too
         onChangeText={setAlarmName}
         value={alarmName}
         placeholder="Alarm"
