+++ components/SoundPicker.js
@@ -1,14 +1,17 @@
 import React from 'react';
 import { View, Text, TouchableOpacity } from 'react-native';
-import {  getDynamicStyles } from '../styles/AlarmStyles';
+import { getDynamicStyles } from '../styles/AlarmStyles';
 import { useDarkMode } from '../contexts/DarkModeContext';
 
-function SoundPicker({ /* pass in any props needed */ }) {
-  const { isDarkMode } = useDarkMode(); 
-  const dynamicStyles = getDynamicStyles(isDarkMode); 
+function SoundPicker(
+  {
+    /* pass in any props needed */
+  }
+) {
+  const { isDarkMode } = useDarkMode();
+  const dynamicStyles = getDynamicStyles(isDarkMode);
 
   return (
-
     <View style={dynamicStyles.container}>
       <Text style={dynamicStyles.label}>Sound</Text>
       <TouchableOpacity
