+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "start": "expo start",
     "tunnel": "expo start --tunnel",
+    "ts:check": "tsc",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
@@ -36,6 +37,11 @@
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
+    "@tsconfig/react-native": "^3.0.3",
+    "@types/jest": "^29.5.12",
+    "@types/react": "^18.2.63",
+    "@types/react-native-vector-icons": "^6.4.18",
+    "@types/react-test-renderer": "^18.0.7",
     "babel-eslint": "^10.1.0",
     "eslint": "^8.56.0",
     "eslint-config-airbnb": "^19.0.4",
@@ -45,7 +51,8 @@
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-native": "^4.1.0",
-    "prettier": "3.2.5"
+    "prettier": "3.2.5",
+    "typescript": "^5.3.3"
   },
   "private": true
 }
